refactor(woozxn) : 오류 수정 및 디렉터리 구조 변경
