+++ src/main/java/com/uninote/dto/response/TokenDto.java
@@ -1,4 +1,4 @@
-package com.uninote.domain;
+package com.uninote.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
