+++ src/main/java/com/uninote/jwt/TokenProvider.java
@@ -1,6 +1,6 @@
 package com.uninote.jwt;
 
-import com.uninote.domain.TokenDto;
+import com.uninote.dto.response.TokenDto;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
